+++ CHANGELOG.md
@@ -5,6 +5,12 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added `confidence` field of type `Option<f32>` to `TiRule` structure
+
 ## [0.11.0] - 2025-06-21
 
 ### Changed
@@ -343,6 +349,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - `client::handshake` implements the application-level handshake process for the
   client after a QUIC connection is established.
 
+[Unreleased]: https://github.com/petabi/review-protocol/compare/0.11.0...main
 [0.11.0]: https://github.com/petabi/review-protocol/compare/0.10.0...0.11.0
 [0.10.0]: https://github.com/petabi/review-protocol/compare/0.9.0...0.10.0
 [0.9.0]: https://github.com/petabi/review-protocol/compare/0.8.1...0.9.0